[Design] 커뮤니티 메인, 커뮤니티 작성, 커뮤니티 신고하기 페이지 UI 구현
